requirements.txt
edgerun-galileo==0.10.4.dev1
galileo-db==0.10.4.dev1

redis~=3.5.3
mysql-connector-python==8.0.16
kubernetes==18.20.0
influxdb_client==1.20.0
python-dotenv